pyproject.toml
[project]
name = "advanced-library-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.2",
    "dj-rest-auth>=7.0.1",
    "django>=5.2",
    "django-allauth>=65.7.0",
    "django-celery-beat>=2.8.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "redis>=6.0.0",
    "requests>=2.32.3",
    "whitenoise>=6.9.0",
]
